apps/flightschool/requirements.txt
# Flight School Application Dependencies
# Note: These specific versions are chosen for stability and compatibility:
# - Flask 3.1.0 and Flask-SQLAlchemy 3.1.1 are maintained for compatibility
# - SQLAlchemy 1.4.51 is used to avoid breaking changes in 2.x
# - Future upgrades should be done with caution and proper testing

# Core Framework
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.1.0
Flask-Mail==0.9.1
python-dotenv==1.0.1
Werkzeug==3.1.0
SQLAlchemy>=2.0.16
WTForms==3.1.2

# Google Calendar Integration
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0

# Validation and Utilities
email-validator==2.1.0.post1
requests
filetype
duckduckgo-search
Pillow
pytz==2024.1
tzlocal==5.0.1

# Development and Testing
pytest==8.0.2
coverage==7.8.0
black==24.2.0
flake8==7.0.0
